refactor(index): split core code base to api, client, log files
